+++ op-node/sources/limit.go
@@ -46,7 +46,6 @@ func (lc *limitClient) EthSubscribe(ctx context.Context, channel any, args ...an
 }
 
 func (lc *limitClient) RawClient() *rpc.Client {
-	// subscription doesn't count towards request limit
 	return lc.c.RawClient()
 }
